requirements.txt
numpy
PyYAML
gymnasium
stable-baselines3
matplotlib
tensorboard